REDCapR: Interaction Between R and REDCap

Encapsulates functions to streamline calls from R to the REDCap API. REDCap (Research Electronic Data CAPture) is a web application for building and managing online surveys and databases developed at Vanderbilt University. The Application Programming Interface (API) offers an avenue to access and modify data programmatically, improving the capacity for literate and reproducible programming.

Comprehensive documentation is also available at https://ouhscbbmc.github.io/REDCapR/.

Much of this package has been developed to support the needs of the following projects. We appreciate the support.

OUHSC CCAN Independent Evaluation of the State of Oklahoma Competitive Maternal, Infant, and Early Childhood Home Visiting (MIECHV) Project. HRSA/ACF D89MC23154. David Bard, PI, OUHSC; 2011-2015.

Independent Evaluation of the State of OK MIECHV Evidence Based Home Visitation Project, NIH-sponsored collaboration with OSDH. David Bard, PI, OUHSC; 2015-2017.

OSDH ParentPRO Pilot Evaluation, federally-sponsored collaboration with OSDH. David Bard, PI, OUHSC; 2015-2017.

Title IV-E Waiver Project, HRSA/MCHB-sponsored collaboration with OKDHS; David Bard, PI, OUHSC; 2014-2017.

Integrative Analysis of Longitudinal Studies of Aging (IALSA), sponsored by NIH 5P01AG043362. Scott Hofer, PI, University of Victoria; Will Beasley, PI of site-award, OUHSC; 2013-2018.

Oklahoma Shared Clinical and Translational Resources, sponsored by NIH NIGMS; U54 GM104938. Judith A. James, PI, OUHSC; 2013-2018.

Additional Institutional Support from OUHSC Dept of Pediatrics; 2013-2017.

Note

The release version is available through CRAN by running install.packages('REDCapR'). The most recent development version is available through GitHub by running remotes::install_github('OuhscBbmc/REDCapR') (make sure remotes is already installed). If you're having trouble with the package, please install the development version. If this doesn't solve your problem, please create a new issue, or email Will.

See REDCapR's advanced vignette for information and examples for overriding the default SSL options.

See also

Author

Maintainer: Will Beasley wibeasley@hotmail.com (ORCID)

Other contributors:

David Bard (ORCID) [contributor]

Thomas Wilson [contributor]

John J Aponte john.aponte@isglobal.org [contributor]

Rollie Parrish rparrish@flightweb.com (ORCID) [contributor]

Benjamin Nutter [contributor]

Andrew Peters (ORCID) [contributor]

Hao Zhu (ORCID) [contributor]

Janosch Linkersdörfer (ORCID) [contributor]

Jonathan Mang (ORCID) [contributor]

Felix Torres fetorres@ucsd.edu [contributor]

Philip Chase pbc@ufl.edu (ORCID) [contributor]

Victor Castro vcastro@mgh.harvard.edu (ORCID) [contributor]

Greg Botwin [contributor]

Stephan Kadauke (ORCID) [contributor]

Ezra Porter (ORCID) [contributor]

Matthew Schuelke matt@themadstatter.com (ORCID) [contributor]

Examples

# \dontrun{
# Install/update REDCapR with the release version from CRAN.
# install.packages('REDCapR')

# Install/update REDCapR with the development version from GitHub
# install.packages("remotes") # Uncomment if `remotes` isn't installed already.
# remotes::install_github('OuhscBbmc/REDCapR')
# }