Enumerate the instruments (forms)

Export a list of the data collection instruments for a project. This includes their unique instrument name as seen in the second column of the Data Dictionary, as well as each instrument's corresponding instrument label, which is seen on a project's left-hand menu when entering data. The instruments will be ordered according to their order in the project. (Copied from "Export Instruments (Data Entry Forms)" method of REDCap API documentation, v.10.5.1)

redcap_instruments(redcap_uri, token, verbose = TRUE, config_options = NULL)

Arguments

redcap_uri

The URI (uniform resource identifier) of the REDCap project. Required.

token

The user-specific string that serves as the password for a project. Required.

verbose

A boolean value indicating if messages should be printed to the R console during the operation. The verbose output might contain sensitive information (e.g. PHI), so turn this off if the output might be visible somewhere public. Optional.

config_options

A list of options to pass to POST method in the httr package. See the details below. Optional.

Value

Currently, a list is returned with the following elements,

data: An R base::data.frame() where each row represents one column in the REDCap dataset.

success: A boolean value indicating if the operation was apparently successful.

status_code: The http status code of the operation.

outcome_message: A human readable string indicating the operation's outcome.

elapsed_seconds: The duration of the function.

raw_text: If an operation is NOT successful, the text returned by REDCap. If an operation is successful, the raw_text is returned as an empty string to save RAM.

Details

The full list of configuration options accepted by the httr package is viewable by executing httr::httr_options(). The httr package and documentation is available at https://cran.r-project.org/package=httr.

References

The official documentation can be found on the 'API Help Page' and 'API Examples' pages on the REDCap wiki (i.e., https://community.projectredcap.org/articles/456/api-documentation.html and https://community.projectredcap.org/articles/462/api-examples.html). If you do not have an account for the wiki, please ask your campus REDCap administrator to send you the static material.

See also

Author

Victor Castro, Will Beasley

Examples

if (FALSE) {
uri           <- "https://bbmc.ouhsc.edu/redcap/api/"
token         <- "9A81268476645C4E5F03428B8AC3AA7B"
ds_instrument <- REDCapR::redcap_instruments(redcap_uri=uri, token=token)$data
}