Creates a dataset that help batching long-running read and writes

The function returns a base::data.frame() that other functions use to separate long-running read and write REDCap calls into multiple, smaller REDCap calls. The goal is to (1) reduce the chance of time-outs, and (2) introduce little breaks between batches so that the server isn't continually tied up.

create_batch_glossary(row_count, batch_size)

Arguments

row_count

The number records in the large dataset, before it's split.

batch_size

The maximum number of subject records a single batch should contain.

Value

Currently, a base::data.frame() is returned with the following columns,

id: an integer that uniquely identifies the batch, starting at 1.

start_index: the index of the first row in the batch. integer.

stop_index: the index of the last row in the batch. integer.

id_pretty: a character representation of id, but padded with zeros.

start_index: a character representation of start_index, but padded with zeros.

stop_index: a character representation of stop_index, but padded with zeros.

label: a character concatenation of id_pretty, start_index, and stop_index_pretty.

Details

This function can also assist splitting and saving a large base::data.frame() to disk as smaller files (such as a .csv). The padded columns allow the OS to sort the batches/files in sequential order.

See also

Author

Will Beasley

Examples

REDCapR::create_batch_glossary(100, 50)
#>   id start_index stop_index index_pretty start_index_pretty stop_index_pretty
#> 1  1           1         50            1                001               050
#> 2  2          51        100            2                051               100
#>       label
#> 1 1_001_050
#> 2 2_051_100
REDCapR::create_batch_glossary(100, 25)
#>   id start_index stop_index index_pretty start_index_pretty stop_index_pretty
#> 1  1           1         25            1                001               025
#> 2  2          26         50            2                026               050
#> 3  3          51         75            3                051               075
#> 4  4          76        100            4                076               100
#>       label
#> 1 1_001_025
#> 2 2_026_050
#> 3 3_051_075
#> 4 4_076_100
REDCapR::create_batch_glossary(100, 3)
#>    id start_index stop_index index_pretty start_index_pretty stop_index_pretty
#> 1   1           1          3           01                001               003
#> 2   2           4          6           02                004               006
#> 3   3           7          9           03                007               009
#> 4   4          10         12           04                010               012
#> 5   5          13         15           05                013               015
#> 6   6          16         18           06                016               018
#> 7   7          19         21           07                019               021
#> 8   8          22         24           08                022               024
#> 9   9          25         27           09                025               027
#> 10 10          28         30           10                028               030
#> 11 11          31         33           11                031               033
#> 12 12          34         36           12                034               036
#> 13 13          37         39           13                037               039
#> 14 14          40         42           14                040               042
#> 15 15          43         45           15                043               045
#> 16 16          46         48           16                046               048
#> 17 17          49         51           17                049               051
#> 18 18          52         54           18                052               054
#> 19 19          55         57           19                055               057
#> 20 20          58         60           20                058               060
#> 21 21          61         63           21                061               063
#> 22 22          64         66           22                064               066
#> 23 23          67         69           23                067               069
#> 24 24          70         72           24                070               072
#> 25 25          73         75           25                073               075
#> 26 26          76         78           26                076               078
#> 27 27          79         81           27                079               081
#> 28 28          82         84           28                082               084
#> 29 29          85         87           29                085               087
#> 30 30          88         90           30                088               090
#> 31 31          91         93           31                091               093
#> 32 32          94         96           32                094               096
#> 33 33          97         99           33                097               099
#> 34 34         100        100           34                100               100
#>         label
#> 1  01_001_003
#> 2  02_004_006
#> 3  03_007_009
#> 4  04_010_012
#> 5  05_013_015
#> 6  06_016_018
#> 7  07_019_021
#> 8  08_022_024
#> 9  09_025_027
#> 10 10_028_030
#> 11 11_031_033
#> 12 12_034_036
#> 13 13_037_039
#> 14 14_040_042
#> 15 15_043_045
#> 16 16_046_048
#> 17 17_049_051
#> 18 18_052_054
#> 19 19_055_057
#> 20 20_058_060
#> 21 21_061_063
#> 22 22_064_066
#> 23 23_067_069
#> 24 24_070_072
#> 25 25_073_075
#> 26 26_076_078
#> 27 27_079_081
#> 28 28_082_084
#> 29 29_085_087
#> 30 30_088_090
#> 31 31_091_093
#> 32 32_094_096
#> 33 33_097_099
#> 34 34_100_100
d <- data.frame(
  record_id = 1:100,
  iv        = sample(x=4, size=100, replace=TRUE),
  dv        = rnorm(n=100)
)
REDCapR::create_batch_glossary(nrow(d), batch_size=40)
#>   id start_index stop_index index_pretty start_index_pretty stop_index_pretty
#> 1  1           1         40            1                001               040
#> 2  2          41         80            2                041               080
#> 3  3          81        100            3                081               100
#>       label
#> 1 1_001_040
#> 2 2_041_080
#> 3 3_081_100