Export the metadata of a REDCap project

Export the metadata (as a data dictionary) of a REDCap project as a base::data.frame(). Each row in the data dictionary corresponds to one field in the project's dataset.

redcap_metadata_read(
  redcap_uri,
  token,
  forms = NULL,
  forms_collapsed = "",
  fields = NULL,
  fields_collapsed = "",
  verbose = TRUE,
  config_options = NULL
)

Arguments

redcap_uri
The URI (uniform resource identifier) of the REDCap project. Required.

token
The user-specific string that serves as the password for a project. Required.

forms
An array, where each element corresponds to the REDCap form of the desired fields. Optional.

forms_collapsed
A single string, where the desired forms are separated by commas. Optional.

fields
An array, where each element corresponds to a desired project field. Optional.

fields_collapsed
A single string, where the desired field names are separated by commas. Optional.

verbose
A boolean value indicating if messages should be printed to the R console during the operation. The verbose output might contain sensitive information (e.g. PHI), so turn this off if the output might be visible somewhere public. Optional.

config_options
A list of options to pass to POST method in the httr package. See the details in redcap_read_oneshot(). Optional.

Value

Currently, a list is returned with the following elements:

data: An R base::data.frame() of the desired records and columns.

success: A boolean value indicating if the operation was apparently successful.

status_codes: A collection of http status codes, separated by semicolons. There is one code for each batch attempted.

outcome_messages: A collection of human readable strings indicating the operations' semicolons. There is one code for each batch attempted. In an unsuccessful operation, it should contain diagnostic information.

forms_collapsed: The desired records IDs, collapsed into a single string, separated by commas.

fields_collapsed: The desired field names, collapsed into a single string, separated by commas.

elapsed_seconds: The duration of the function.

Details

Specifically, it internally uses multiple calls to redcap_read_oneshot() to select and return data. Initially, only primary key is queried through the REDCap API. The long list is then subsetted into partitions, whose sizes are determined by the batch_size parameter. REDCap is then queried for all variables of the subset's subjects. This is repeated for each subset, before returning a unified base::data.frame().

The function allows a delay between calls, which allows the server to attend to other users' requests.

References

The official documentation can be found on the 'API Help Page' and 'API Examples' pages on the REDCap wiki (i.e., https://community.projectredcap.org/articles/456/api-documentation.html and https://community.projectredcap.org/articles/462/api-examples.html). If you do not have an account for the wiki, please ask your campus REDCap administrator to send you the static material.

Author

Will Beasley

Examples

if (FALSE) {
uri   <- "https://bbmc.ouhsc.edu/redcap/api/"
token <- "9A81268476645C4E5F03428B8AC3AA7B"
REDCapR::redcap_metadata_read(redcap_uri=uri, token=token)
}