List authorized users

List users authorized for a project.

redcap_users_export(redcap_uri, token, verbose = TRUE, config_options = NULL)

Arguments

redcap_uri
The URI (uniform resource identifier) of the REDCap project. Required.

token
The user-specific string that serves as the password for a project. Required.

verbose
A boolean value indicating if messages should be printed to the R console during the operation. The verbose output might contain sensitive information (e.g. PHI), so turn this off if the output might be visible somewhere public. Optional.

config_options
A list of options to pass to POST method in the httr package. See the details below. Optional.

Value

a utils::packageDescription.

Note

Documentation in REDCap 8.4.0

This method allows you to export the list of users for a project,
including their user privileges and also email address, first name,
and last name.

Note: If the user has been assigned to a user role, it will return
the user with the role's defined privileges.

Examples

uri      <- "https://bbmc.ouhsc.edu/redcap/api/"
token    <- "06DEFB601F9B46847DAA9DF0CFA951B4"
result   <- REDCapR::redcap_users_export(redcap_uri=uri, token=token)

#> The REDCap users were successfully exported in 0.3 seconds.  The http status code was 200.

result$data_user

#> # A tibble: 2 × 31
#>   username        email firstname lastname expiration data_access_gro… data_access_gro…
#>   <chr>           <chr> <chr>     <chr>    <date>     <chr>            <chr>           
#> 1 unittestphifree wibe… Unit Test PHI Free NA         daga             331             
#> 2 wbeasleya       will… Will      Beasley… NA         NA               NA              
#> # … with 24 more variables: design <lgl>, user_rights <lgl>,
#> #   data_access_groups <lgl>, data_export <chr>, reports <lgl>,
#> #   stats_and_charts <lgl>, manage_survey_participants <lgl>, calendar <lgl>,
#> #   data_import_tool <lgl>, data_comparison_tool <lgl>, logging <lgl>,
#> #   file_repository <lgl>, data_quality_create <lgl>,
#> #   data_quality_execute <lgl>, api_export <lgl>, api_import <lgl>,
#> #   mobile_app <lgl>, mobile_app_download_data <lgl>, record_create <lgl>, …

result$data_user_form

#> # A tibble: 2 × 4
#>   username        form_name    permission_id permission
#>   <chr>           <chr>                <int> <fct>     
#> 1 unittestphifree demographics             1 edit_form 
#> 2 wbeasleya       demographics             1 edit_form