Determine version of REDCap instance

This function uses REDCap's API to query its version.

redcap_version(redcap_uri, token, verbose = TRUE, config_options = NULL)

Arguments

redcap_uri
The URI (uniform resource identifier) of the REDCap project. Required.

token
The user-specific string that serves as the password for a project. Required.

verbose
A boolean value indicating if messages should be printed to the R console during the operation. The verbose output might contain sensitive information (e.g. PHI), so turn this off if the output might be visible somewhere public. Optional.

config_options
A list of options to pass to POST method in the httr package. See the details below. Optional.

Value

a utils::packageDescription

Details

If the API call is unsuccessful, a value of base::package_version("0.0.0") will be returned. This ensures that a the function will always return an object of class base::numeric_version. It guarantees the value can always be used in utils::compareVersion().

Examples

uri      <- "https://bbmc.ouhsc.edu/redcap/api/"
token    <- "9A81268476645C4E5F03428B8AC3AA7B"
REDCapR::redcap_version(redcap_uri = uri, token = token)

#> The REDCap version was successfully determined in 0.3 seconds. The http status code was 200.  Is is 10.5.1.

#> [1] ‘10.5.1’